aligned refresh in promise cache to lazy promise
